requirements.txt
docker>=7.1.0
requests>=2.32.3
openai>=1.64.0
huggingface-hub>=0.29.1
groq>=0.23.1
tree-sitter>=0.24.0
tree-sitter-javascript>=0.23.1
python-dotenv>=1.1.0
pytest>=8.4.1